backend/requirements.txt
requests
cryptography
python-dotenv
websockets
streamlit
websocket-client
rel

# Core scientific stack used by backend/finfill.py
numpy>=1.20
scikit-learn>=1.2
joblib>=1.2
pandas>=1.3

# Deep learning for GRU model
tensorflow>=2.13.0
